Working on: Stocks Entity

+++ routes/api.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\ProductsController;
+use App\Http\Controllers\StocksController;
 use App\Http\Controllers\StoreController;
 use App\Http\Controllers\SupplierController;
 use Illuminate\Http\Request;
@@ -38,46 +39,45 @@ Route::group([
     }
 );
 
-
 /**
  * Route group for auth controller.
  */
 Route::controller(AuthController::class)
     ->prefix('/users')
-    ->group(function () {
-        Route::get('roles', 'allAvailableRoles');
+    ->group(function(){
+        Route::get('roles', 'allAvailableRoles')->middleware(['auth.anyrole:ROLE_ADMIN,ROLE_STORE_OWNER']);
         Route::get('logout', 'logout');
         Route::get('current-user', 'getUser');
-        Route::get('all-users', 'getAllUser');
+        Route::get('all-users', 'getAllUser')->middleware(['auth.anyrole:ROLE_ADMIN,ROLE_STORE_OWNER']);
         Route::post('login', 'login');
-        Route::post('register', 'register');
+        Route::post('register', 'register')->middleware(['auth.anyrole:ROLE_ADMIN,ROLE_STORE_OWNER']);
         Route::post('refresh', 'refresh');
         Route::post('has-permission', 'hasRole');
         Route::post('has-any-roles', 'hasAnyRole');
         Route::post('has-all-roles', 'hasAllRoles');
         Route::patch('update-user-profile', 'editUser');
-        Route::delete('delete-user/{user}', 'deleteUser');
+        Route::delete('delete-user/{user}', 'deleteUser')->middleware(['auth.anyrole:ROLE_ADMIN,ROLE_STORE_OWNER']);
     });
 
 /**
- * Route group for store interaction
+ * Route group for store interactions
  */
 Route::controller(StoreController::class)
     ->prefix('/stores')
-    ->group(function () {
-        Route::get('all-stores', 'index');
-        Route::get('store/{store}', 'show');
-        Route::post('add-store', 'store');
-        Route::patch('update-store/{store}', 'update');
-        Route::delete('delete-store/{store}', 'destroy');
+    ->group(function(){
+        Route::get('all-stores', 'index')->middleware('auth.allrole:ROLE_ADMIN');
+        Route::get('store/{store}', 'show')->middleware('auth.anyrole:ROLE_ADMIN,ROLE_STORE_OWNER');
+        Route::post('add-store', 'store')->middleware('auth.allrole:ROLE_ADMIN');
+        Route::patch('update-store/{store}', 'update')->middleware('auth.allrole:ROLE_ADMIN');
+        Route::delete('delete-store/{store}', 'destroy')->middleware('auth.allrole:ROLE_ADMIN');
     });
 
 /**
- * Route group for supplier interaction
+ * Route group for supplier interactions
  */
 Route::controller(SupplierController::class)
     ->prefix('/suppliers')
-    ->group(function () {
+    ->group(function(){
         Route::get('all-suppliers', 'index');
         Route::get('supplier/{supplier}', 'show');
         Route::post('add-supplier', 'store');
@@ -85,9 +85,12 @@ Route::controller(SupplierController::class)
         Route::delete('delete-supplier/{supplier}', 'destroy');
     });
 
+/**
+ * Route group for products interactions
+ */
 Route::controller(ProductsController::class)
     ->prefix('/products')
-    ->group(function () {
+    ->group(function(){
         Route::get('all-products', 'index');
         Route::get('product/{product}', 'show');
         Route::post('add-product', 'store');
@@ -95,4 +98,14 @@ Route::controller(ProductsController::class)
         Route::put('add-payable/{product}', 'add_payable');
         Route::put('add-paid/{product}', 'add_paid');
         Route::delete('delete-product/{product}', 'destroy');
+    });
+
+/**
+ * Route group for stocks interactions
+ */
+Route::controller(StocksController::class)
+    ->prefix('/stocks')
+    ->group(function(){
+        Route::get('all-stocks/{store}', 'index');
+
     });
